requirements (1).txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
PyJWT==2.7.0
requests==2.31.0